[refactor] di를 app 모듈단으로 이동시킨다.
